feat(fdc): preliminary implementation of INT signal
INT is used to notify asynchronous events to the MCU (i.e. seek
command is sent, INT rises when the head has moved onto the requested
track). This is used too to notify if new data burst is available during
execution phase.

If no disk is present and read command is requested, INT=0.
In this implementation, a disk inserted after the read command, will
trigger INT and it should unlock the reading process.