fix(fdc): read/write CHS result generation in error condition are now compliant
Introduced IDR (ID Register) and next IDR to handle the result phase in
both success and error condition.
Briefly, when one or more sectors are succesfully read, IDR is
incremented to the next one (handling head/track overflow).
If an error occur, the CHS in result are referred to the latest
correctly read sector.
Tested on the actual FDC device.