Return NULL when tokenizer input is NULL.
This commit allows the user to pass NULL to the source of the
tokenizer_next_hex() function. This is not considered an error anymore.
Instead, passing NULL just does nothing and returns NULL.
This behaviour is useful to write more compact parsers and avoid
bloating the code with several `if...else` condition inside the command
parser.
With this commit, an user can pass around the same `src`/`dst` pointer
multiple times, and only check which calls have failed at the end.